refactor: 사용자관련 리소스 조회 라우팅 구조 수정
